refactoring: moved git functions to own class
