+++ src/ai/susi/server/api/cms/DeleteSkillService.java
@@ -1,6 +1,7 @@
 package ai.susi.server.api.cms;
 
 import ai.susi.DAO;
+import ai.susi.SkillTransactions;
 import ai.susi.json.JsonObjectWithDefault;
 import ai.susi.server.*;
 import org.eclipse.jgit.api.Git;
@@ -115,26 +116,26 @@ public class DeleteSkillService extends AbstractAPIHandler implements APIHandler
             //Add to git
             if (privateSkill != null) {
                 DAO.deleteChatbot(userId, group_name, language_name, skill_name);
-                try (Git git = DAO.getPrivateGit()) {
+                try (Git git = SkillTransactions.getPrivateGit()) {
                     git.add()
                     .setUpdate(true)
                     .addFilepattern(".")
                     .call();
                 // and then commit the changes
-                    DAO.pushCommit(git, "Deleted " + skill_name, !rights.getIdentity().isAnonymous() ? rights.getIdentity().getName() : "anonymous@");
+                    SkillTransactions.pushCommit(git, "Deleted " + skill_name, !rights.getIdentity().isAnonymous() ? rights.getIdentity().getName() : "anonymous@");
                     json.put("message", "Deleted " + skill_name);
                 } catch (IOException | GitAPIException e) {
                     e.printStackTrace();
                 }
             }
             else {
-                try (Git git = DAO.getGit()) {
+                try (Git git = SkillTransactions.getGit()) {
                     git.add()
                     .setUpdate(true)
                     .addFilepattern(".")
                     .call();
                 // and then commit the changes
-                    DAO.pushCommit(git, "Deleted " + skill_name, !rights.getIdentity().isAnonymous() ? rights.getIdentity().getName() : "anonymous@");
+                    SkillTransactions.pushCommit(git, "Deleted " + skill_name, !rights.getIdentity().isAnonymous() ? rights.getIdentity().getName() : "anonymous@");
                     json.put("message", "Deleted " + skill_name);
                 } catch (IOException | GitAPIException e) {
                     e.printStackTrace();
